+++ lib/cldr/calendar/calendars/julian.ex
@@ -18,6 +18,9 @@ defmodule Cldr.Calendar.Julian do
       @new_year_starting_month start_month
       @new_year_starting_day start_day
 
+      @months_in_year Cldr.Calendar.Julian.months_in_year(0)
+      @last_month_of_year rem(start_month + (@months_in_year - 1), @months_in_year)
+
       # Adjust the year to be a Jan 1st starting year and carry
       # on
       def date_to_iso_days(year, month, day) when month <= @new_year_starting_month and day < @new_year_starting_day do
@@ -26,6 +29,10 @@ defmodule Cldr.Calendar.Julian do
 
       def date_to_iso_days(year, month, day) do
         Cldr.Calendar.Julian.date_to_iso_days(year, month, day)
+      end
+
+      def naive_datetime_to_iso_days(year, month, day, 0, 0, 0, {0, 0}) do
+        {date_to_iso_days(year, month, day), {0, 0}}
       end
 
       # Adjust the year to be this calendars starting year
@@ -39,22 +46,78 @@ defmodule Cldr.Calendar.Julian do
         end
       end
 
+      # Here we use month to mean ordinal month. Therefore if the calendar
+      # starts on March 25th, then days in month for March will be seen as
+      # days if month for month 1.
+
+      def days_in_month(year, ordinal_month) do
+        adjusted_month =
+          Cldr.Math.amod(ordinal_month + @new_year_starting_month - 1, @months_in_year)
+
+        adjusted_year =
+          if adjusted_month < @new_year_starting_month, do: year + 1, else: year
+
+        cond do
+          # The first month of the year will be short since the year starts
+          # part way through the month
+          adjusted_month == @new_year_starting_month ->
+            days_in_julian_month = Cldr.Calendar.Julian.days_in_month(year, ordinal_month)
+            days_in_julian_month - @new_year_starting_day
+
+          # The last month of the year will be "long" since the first part of the
+          # first month that is before the start of year will be included
+          adjusted_month == @last_month_of_year ->
+            start_of_month =
+              date_to_iso_days(adjusted_year, adjusted_month, 1)
+            start_of_next_month =
+              date_to_iso_days(adjusted_year, @new_year_starting_month, @new_year_starting_day)
+            start_of_next_month - start_of_month
+
+          true ->
+            Cldr.Calendar.Julian.days_in_month(year, adjusted_month)
+        end
+      end
+
+      def month(year, ordinal_month) do
+        adjusted_month =
+          Cldr.Math.amod(ordinal_month + @new_year_starting_month - 1, @months_in_year)
+
+        adjusted_year =
+          if adjusted_month < @new_year_starting_month, do: year + 1, else: year
+
+        first_day =
+          if adjusted_month == @new_year_starting_month, do: @new_year_starting_day, else: 1
+
+        {:ok, first} = Date.new(adjusted_year, adjusted_month, first_day, __MODULE__)
+        first_iso_days = date_to_iso_days(year, adjusted_month, first_day)
+        days_in_month = days_in_month(year, ordinal_month)
+
+        last_iso_days = first_iso_days + days_in_month - 1
+        {year, month, day} = date_from_iso_days(last_iso_days)
+        {:ok, last} = Date.new(year, month, day, __MODULE__)
+
+        Date.range(first, last, 1)
+      end
+
+      # Returns the ordinal month
+      def month_of_year(_year, ordinal_month, _day) do
+        Cldr.Math.amod(ordinal_month - @new_year_starting_month + 1, @months_in_year)
+      end
+
       defdelegate valid_date?(year, month, day), to: Cldr.Calendar.Julian
       defdelegate leap_year?(year), to: Cldr.Calendar.Julian
       defdelegate plus(year, month, day, part, years, options), to: Cldr.Calendar.Julian
       defdelegate week(year, week), to: Cldr.Calendar.Julian
-      defdelegate month(year, month), to: Cldr.Calendar.Julian
       defdelegate weeks_in_year(year), to: Cldr.Calendar.Julian
-      defdelegate days_in_month(year, month), to: Cldr.Calendar.Julian
       defdelegate day_of_week(year, month, day, starts_on), to: Cldr.Calendar.Julian
       defdelegate day_of_year(year, month, day), to: Cldr.Calendar.Julian
       defdelegate day_of_era(year, month, day), to: Cldr.Calendar.Julian
       defdelegate iso_week_of_year(year, month, day), to: Cldr.Calendar.Julian
       defdelegate week_of_year(year, month, day), to: Cldr.Calendar.Julian
-      defdelegate month_of_year(year, month, day), to: Cldr.Calendar.Julian
       defdelegate quarter_of_year(year, month, day), to: Cldr.Calendar.Julian
       defdelegate year_of_era(year), to: Cldr.Calendar.Julian
       defdelegate parse_date(date), to: Cldr.Calendar.Julian
+      defdelegate date_to_string(year, month, day), to: Cldr.Calendar.Julian
     end
   end